+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ChangeLog
 =====================
 
+1.2.2
+=====================
+* #283 Send thumbs-down icon when build fails
+* #284 NullPointerException when using the GitLabCommitStatusPublisher
+* Use jenkins credentials-plugin for storing the GitLab API Token
+* Clear clients cache if connection config has changed
+* Add workflow step that updates the GitLab commit status depending on the build status
+
 1.2.1
 =====================
 * #271 Version 1.2.0 breaks support for Workflow/Pipeline jobs